Move startup code to separate files Convert DataHubProvider to functional component
